requirements.txt
pandas
numpy
streamlit
openai
langchain
transformers
langchain-openai
python-dotenv
torch
langchain-community
openpyxl
